xxd.c and libxxd.c are now seperate
